+++ Immortal/Platform/D3D12/Buffer.cpp
@@ -104,7 +104,7 @@ void Buffer::__Create()
 #endif
 }
 
-void Buffer::Update(uint32_t updateSize, const void *data, uint32_t offset)
+void Buffer::Update(uint64_t updateSize, const void *data, uint64_t offset)
 {
     THROWIF(updateSize > size, SError::OutOfBound);
